requirements.txt
torch>=2.1
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
tqdm>=4.66
PyYAML>=6.0
tensorboard>=2.15
psutil>=5.9
